requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
cffi==1.15.1
    # via cryptography
colorama==0.4.6
    # via rsarmageddon
cryptography==39.0.2
    # via -r requirements.in
gmpy2==2.1.5
    # via rsarmageddon
psutil==5.9.4
    # via rsarmageddon
pycparser==2.21
    # via cffi
pycryptodome==3.17
    # via rsarmageddon
pycryptodomex==3.17
    # via -r requirements.in
rsarmageddon==2.1.1
    # via -r requirements.in
